Add the infra needed for the remote state, and the remote state configuration itself
The configuration that tells terraform to use S3 as the storage of the
terraform state (so the terraform.tfstate file can be distributed
among the team mates) is done in the terraform.tf file.
This configuration needs bucket and a dynamoDB table (to lock the file
while updating it), those things are created in the setup/setup.tf file.
A script was added to first create the infra and then run the configuration
of the remote state (I'm not sure there's another way to do this)